shifted odeprepare to new file
